add Bader's draft about OOP in Fortran to the learning section (#477)
* add Bader's draft about OOP features in Fortran

The draft by the late Reinhold Bader (1966-2024) about OOP features
in Fortran is added to the learning section.  Except the addition to
the `learning.yml` file and a boiler plate header in the mandatory
`index.md` file, it is a copy-paste of the illustrations and the
text provided by `pandoc_md.md`, commit a77670d by 2024-09-03 of
the rescuing GitHub repository.[1]

[1] https://github.com/nbehrnd/bader_article

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* explicitly credit Reinhold Bader as author

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* test Sphinx' tip environment

The contribution page[1] describes optional admonition environments
which could help to retrieve hints.  It is tested on one case.

[1] https://fortran-lang.org/en/community/contributing/

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* extend usage of admonitions

Now each instance of string `_Hint:_` is substituted by an
admonition environment.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* adjust indentations to 2 spaces/level

Snippets of code in the learning section indent with 2 spaces per
level eachl.  This pattern now is used, mostly with help of findent.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* revise usage of f90 code block fence label

Following the requested Code style,[1] snippets of code which show
a principe but as such can not be successfully compiled no longer
carry the `f90` label.  The presence of at least one colon `:`,
or one elipsis `...` is considered as an indicator.

[1] https://fortran-lang.org/en/community/contributing/

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* redefine images as block level images

The usage of the two illustrations were edited; they now are block
level images.[1]

[1] https://myst-parser.readthedocs.io/en/latest/syntax/images_and_figures.html

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* change Fortran keywords to lower-case letters

Following a recommendation,[1] keywords of Fortran in the snippets
of code, running text, and section headers now are set in lower-
case characters only.  Beside syntax highlighlighting the web site
offers, this then is more consistent in style with the other booklets
already present.  There however are exceptions to this change in the
table about value statements, section Performing I/O with objects of
container-like type).

[1] https://github.com/fortran-lang/webpage/pull/477

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* add booklet to the index of the learning section

Bader's draft about the support of OOP by Fortran now equally is
accessible from the centre (section Fortran documentation) as well
as from the lateral menu of the booklets.  The entry in the later
is new by this commit.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* split oop single page booklet into a multipage booklett

The booklet's content now is organized in three pages.  This deviates
from the state of commit a77670d (2024-09-03) of the rescuing GitHub
respository,[1] however likely reflects better the original intent by
Bader.

[1] https://github.com/nbehrnd/bader_article

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* lightly lint the markdown file for eventual aesthetics

A light edit to improve the aesthetics of the eventual rendering
as html file.  The cross-links within the booklet are removed because
I don't understand the working of internal site links as explained by
the guidelines,[1] nor successfully identified a working example among
the other booklets by others.

[1] https://fortran-lang.org/en/community/contributing/

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

* Update source/learn/oop_features_in_fortran/object_based_programming_techniques.md

Co-authored-by: Federico Perini <federico.perini@gmail.com>

* Update source/learn/oop_features_in_fortran/object_based_programming_techniques.md

Co-authored-by: Federico Perini <federico.perini@gmail.com>

* Update performance_and_ease_of_use.md

add code fence labels `f90`

* add missing f90 code fence labels

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>

---------

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>
Co-authored-by: Federico Perini <federico.perini@gmail.com>