feat(f95_features.md): initiate work
Based on a request on fortran-lang.org,[1] work for a new
booklet for the learning section of fortran-lang.org about
the features of Fortran 95 starts.

The initial step was to fetch the source code (mediawiki synatax)
which was converted with Pandoc to GitHub flavored Markdown where
possible.  This was not possible for the trailing bibliography
section of the Wikipedia article (hence left as such in a fenced
code block), which however is more an optional "for an additional
read, consult ..." rather than a specific literature reference to
a particular keyword, or phrase.

While GitHub flavored Markdown isn't (yet) MyST Markdown flavor
required, it is close enough for the manual edit ahed.  Equally
a TODO is to build a multipage booklet, i.e. an `index.md` and
individual Markdown files about each chapter; this gradually
removes `f95_features.md` as an intermediate container file.

[1] https://fortran-lang.discourse.group/t/fortran-95-language-features/9263

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>